requirements.txt
newsapi-python>=0.2.7
plotly>=5.0.0
python-dotenv>=1.0.0
streamlit>=1.35.0
torch>=2.0.0
transformers>=4.40.0
